Relocate common fixtures
