+++ exporter/f680/application_sections/conftest.py
@@ -1,0 +1,38 @@
+import pytest
+
+from core import client
+
+
+@pytest.fixture()
+def unset_f680_feature_flag(settings):
+    settings.FEATURE_FLAG_ALLOW_F680 = False
+
+
+@pytest.fixture(autouse=True)
+def setup(mock_exporter_user_me, settings):
+    settings.FEATURE_FLAG_ALLOW_F680 = True
+
+
+@pytest.fixture
+def missing_application_id():
+    return "6bb0828c-1520-4624-b729-7f3e6e5b9f5d"
+
+
+@pytest.fixture
+def mock_f680_application_get_404(requests_mock, missing_application_id):
+    url = client._build_absolute_uri(f"/exporter/f680/application/{missing_application_id}/")
+    return requests_mock.get(url=url, json={}, status_code=404)
+
+
+@pytest.fixture
+def mock_f680_application_get(requests_mock, data_f680_case):
+    application_id = data_f680_case["id"]
+    url = client._build_absolute_uri(f"/exporter/f680/application/{application_id}/")
+    return requests_mock.get(url=url, json=data_f680_case)
+
+
+@pytest.fixture
+def mock_patch_f680_application(requests_mock, data_f680_case):
+    application_id = data_f680_case["id"]
+    url = client._build_absolute_uri(f"/exporter/f680/application/{application_id}/")
+    return requests_mock.patch(url=url, json=data_f680_case)
